M3 mutable label box size now set with float rather than int

+++ multiplicity3-jme3contentsystem/src/main/java/multiplicity3/jme3csys/items/mutablelabel/JMEMutableLabelDelegate.java
@@ -112,7 +112,7 @@ public class JMEMutableLabelDelegate extends JMEItemDelegate implements
 	}
 
 	@Override
-	public void setBoxSize(int width, int height) {
+	public void setBoxSize(float width, float height) {
 		boundingBox = new Rectangle(0, 0, width, height);
 		doUpdate();
 	}
